refactor: reorganize minio config and implement security propagation in consumer
- Relocate MinIO configuration classes to 'config.minio' package
- Implement security context propagation in MessageConsumer using 'X-ORIGIN-USER-ID' header
- Remove legacy file importer classes and factory logic
- Clean up unused imports and debug logging in application.yaml
