+++ src/main/java/br/com/ifrn/AcademicService/AcademicServiceApplication.java
@@ -1,12 +1,11 @@
 package br.com.ifrn.AcademicService;
 
 import br.com.ifrn.AcademicService.config.redis.RedisPropertiesConfig;
-import br.com.ifrn.AcademicService.file.objectstorage.MinioPropertiesConfig;
+import br.com.ifrn.AcademicService.config.minio.MinioPropertiesConfig;
 import br.com.ifrn.AcademicService.config.keycloak.KeycloakPropertiesConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.web.bind.annotation.GetMapping;
 
 @EnableConfigurationProperties({
         KeycloakPropertiesConfig.class,
